Move the pyPhotometry edge cases off the gin data
Six of them wrote their own bytes into tmp_path and read no recording, so they sat in
tests/test_on_data and only ran when NEUROCONV_TEST_DATA_PATH was set. They cover the
paths most likely to rot unnoticed, as each one is a case no recording covers: an
unknown mode, an unknown legacy mode code, a version 1.1 layout, a blank subject id, a
subject id filling its fixed field, and a header readable as neither generation.

The two tests that did open a published recording move into the round-trip class for
the file they were already using, so the mixed class goes away.
